docker-image/requirements.txt
fastapi
uvicorn
Pillow
ultralytics
gradio
pydantic
numpy
matplotlib